frontend-tools: Move source helper functions to a header
